Moved test to own test case
